Fold CsprojRewriter into RewriteCsprojsStep
CsprojRewriter.Rewrite had a single caller (RewriteCsprojsStep). Move its
regex fields and Rewrite method directly into the step and delete the
now-empty class. Rename the spec file/class to match.
